requirements.txt
pymupdf
Pillow